pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "underdogcowboy"
version = "0.1.14.13"
description = "Underdog Cowboy (UC): Wrangle Your LLMs with a Smile"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}

dependencies = [
    "vertexai",
    "flask",
    "aiohttp",
    "aiofiles",
    "prompt_toolkit", 
    "python-dotenv",
    "keyring",
    "groq",
    "langsmith",
    "reportlab",
    "pillow",
    "requests-mock",
    "rich",
    "mdformat",
    "textual",
    "pyyaml",
    "pyperclip",
    "uccli>=0.0.5.1",
    "pywin32; platform_system == 'Windows'"
]

[project.scripts]
timelineeditor = "underdogcowboy.run_timeline_editor:main"
uc-llm = "underdogcowboy.core.commandtools.llm_config:main"
uc-poke = "underdogcowboy.core.commandtools.poke:main"
uc-timelineeditor = "underdogcowboy.run_timeline_editor:main"
uc-tracing = "underdogcowboy.core.commandtools.tracing_config:main"
uc-agent-clarity =  "underdogcowboy.core.commandtools.agent_clarity:main"
uc-agent-pdf = "underdogcowboy.core.commandtools.agent_pdf:main"
uc-assessment-builder = "underdogcowboy.core.commandtools.assessment_agent_builder:main"
uc-agent-flow = "underdogcowboy.core.commandtools.agent_flow.main:main"
ucb-agents = "underdogcowboy.flask_apps.agent_manager.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["underdogcowboy*"]

[tool.setuptools.package-data]
"underdogcowboy.agents" = ["*.json"]
"underdogcowboy.flask_apps.agent_manager.react_ui" = ["static/*", "index.html"]
